+++ modules/play/src/main/scala/io/bartholomews/fsclient/play/FsClientPlayApi.scala
@@ -21,7 +21,7 @@ trait FsClientPlayApi {
   private val tokenCommonFieldsJsonPath =
     (JsPath \ "generated_at")
       .read[Long]
-      .orElse(Reads.pure(System.currentTimeMillis))
+      .orElse(Reads.pure(System.nanoTime()))
       .and((JsPath \ "access_token").read[AccessToken])
       .and((JsPath \ "token_type").read[String])
       .and((JsPath \ "expires_in").read[Long])